+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Add recommendations for replacement plugins related to @ionic-enterprise
 - Show Capacitor recommendations even if cordova remnants are in the project
 - Hide disconnected devices when running iOS/Android
+- Expand native project properties by default
 
 ### Version 1.22.1